requirements.txt
Flask>=1.0
psycopg2-binary>=2.7
gunicorn>=19.9
passlib>=1.6.1
WTForms>=2.2.1
